chat/requirements.txt
openai                    == 0.27.8
python-dotenv             == 1.0.0

# app
gradio                    == 3.35.0